refactor: move compat layer to separate package
