+++ package.json
@@ -15,6 +15,8 @@
   ],
   "scripts": {
     "test": "jest",
+    "lint": "eslint .",
+    "coveralls": "jest --coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "integration": "jest --config jest.integration.config.json --setupTestFrameworkScriptFile=./jest.integration.setup.js"
   },
   "repository": {
