+++ packages/next-aws-lambda/lib/__tests__/compatLayer.response.test.js
