+++ packages/next-aws-lambda/lib/compatLayer.js
