+++ packages/next-aws-lambda/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "next-aws-lambda",
+  "version": "1.8.1",
+  "description": "Compat layer between next.js serverless page and AWS Lambda",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/danielcondemarin/serverless-nextjs-plugin.git"
+  },
+  "author": "Daniel Conde Marin <danielconde9@gmail.com>",
+  "license": "ISC"
+}
