+++ packages/next-plugin/lib/__tests__/getFactoryHandlerCode.test.js
@@ -7,27 +7,21 @@ describe("getFactoryHandlerCode", () => {
     const compatHandlerContent = getFactoryHandlerCode(
       path.join(PluginBuildDir.BUILD_DIR_NAME, "my-page.js")
     );
-    expect(compatHandlerContent).toContain(
-      'require("serverless-nextjs-plugin/aws-lambda-compat")'
-    );
+    expect(compatHandlerContent).toContain('require("next-aws-lambda")');
   });
 
   it("should require compatLayer with correct path when page is nested", () => {
     const compatHandlerContent = getFactoryHandlerCode(
       path.join(PluginBuildDir.BUILD_DIR_NAME, "categories/fridge/fridges.js")
     );
-    expect(compatHandlerContent).toContain(
-      'require("serverless-nextjs-plugin/aws-lambda-compat")'
-    );
+    expect(compatHandlerContent).toContain('require("next-aws-lambda")');
   });
 
   it("should require compatLayer with correct path when buildDir is nested", () => {
     const compatHandlerContent = getFactoryHandlerCode(
       path.join(`app/${PluginBuildDir.BUILD_DIR_NAME}`, "page.js")
     );
-    expect(compatHandlerContent).toContain(
-      'require("serverless-nextjs-plugin/aws-lambda-compat")'
-    );
+    expect(compatHandlerContent).toContain('require("next-aws-lambda")');
   });
 
   it("should require next page provided", () => {
